% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSINS-package.R
\docType{package}
\name{rSINS-package}
\alias{rSINS}
\alias{rSINS-package}
\title{rSINS: SINS data in R}
\description{
Loads data generated from SINS into R.
}
\author{
\strong{Maintainer}: Tiago F.V. Maié \email{tiagomaie@hotmail.com}

}
\keyword{internal}
